requirements.txt
# or opeani agent
openai
openai-agents
# for api extraction
python-dotenv
# frontend
streamlit
# for the calendar agent
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client